UI/requirements.txt
Flask == 2.3.2
Elasticsearch == 8.8.0
python-dotenv 
Flask-Login==0.5.3
Flask-SQLAlchemy==2.5.1
Flask-Bcrypt==0.7.1
pycountry==20.7.3
flask-login==0.5.3
flask_sqlalchemy==2.5.1
flask_bcrypt==0.7.1
 